% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise_dist.R
\name{optimise_distance}
\alias{optimise_distance}
\title{R package to analyse univariate distributions}
\usage{
optimise_distance(dataset, method, k)
}
\arguments{
\item{dataset}{A data matrix or data frame.}

\item{method}{If package is supposed to treat input dataset as univariate distributions then method name should be "pdf" else "non-pdf" to treat this dataset as continuous data vectors.}

\item{k}{An integer value used as to feed number of clusters.}
}
\value{
cluster groups of univariate distributions
}
\description{
R package to analyse univariate distributions
}
\section{Details}{

This package offers a flexibility to find an optimised distance measure on the basis of provided dataset for both continuous data vectors as well as for univariate distributions computed through kernel density function. The performance of optimized distance measure is evaluated through analyzing the clustering results.
}

\examples{
\dontrun{
optimise_distance(dataset, method = "pdf", k = 3)
}

}
